+++ sources/put_img_to_img.c
@@ -6,7 +6,7 @@
 /*   By: deydoux <deydoux@student.42lyon.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/28 16:23:03 by deydoux           #+#    #+#             */
-/*   Updated: 2024/03/14 15:21:48 by deydoux          ###   ########.fr       */
+/*   Updated: 2024/03/14 15:28:35 by deydoux          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,7 @@ void	put_img_to_img(t_img src, t_img dst, size_t x, size_t y)
 	x *= 4;
 	width = src.width * 4;
 	y_src = 0;
-	while (y_src < (size_t)src.heigh)
+	while (y_src < src.heigh)
 	{
 		x_src = 0;
 		while (x_src < width)
